Extract github section
Move into re-useable module
